:recycle: Refactor raw gradient data into wasm module
